backend_service/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-firestore
google-cloud-storage
python-multipart
